tree: bulk update

+++ files/initrd/opt/arc/zimage-patch.sh
@@ -4,12 +4,13 @@
 
 . "${ARC_PATH}/include/functions.sh"
 
+arc_mode
+
 set -o pipefail # Get exit code from process piped
 
 # Sanity check
 [ -f "${ORI_ZIMAGE_FILE}" ] || (die "${ORI_ZIMAGE_FILE} not found!" | tee -a "${LOG_FILE}")
-
-rm -f "${MOD_ZIMAGE_FILE}"
+[ -f "${MOD_ZIMAGE_FILE}" ] && rm -f "${MOD_ZIMAGE_FILE}"
 
 KERNEL="$(readConfigKey "kernel" "${USER_CONFIG_FILE}")"
 PLATFORM="$(readConfigKey "platform" "${USER_CONFIG_FILE}")"
@@ -17,11 +18,11 @@ PRODUCTVER="$(readConfigKey "productver" "${USER_CONFIG_FILE}")"
 KVER="$(readConfigKey "platforms.${PLATFORM}.productvers.\"${PRODUCTVER}\".kver" "${P_FILE}")"
 is_in_array "${PLATFORM}" "${KVER5L[@]}" && KVERP="${PRODUCTVER}-${KVER}" || true
 if [ "${KERNEL}" = "custom" ]; then
-  echo -e ">> using customized Kernel for ${PLATFORM} (${KVER})"
+  [ "${ARC_MODE}" != "dsm" ] && echo -e ">> Kernel: customized ${PLATFORM} (${KVER})"
   # Extract bzImage
   gzip -dc "${CUSTOM_PATH}/bzImage-${PLATFORM}-${KVERP}.gz" >"${MOD_ZIMAGE_FILE}"
 else
-  echo -e ">> using official Kernel for ${PLATFORM} (${KVER})"
+  [ "${ARC_MODE}" != "dsm" ] && echo -e ">> Kernel: official ${PLATFORM} (${KVER})"
   # Extract vmlinux
   "${ARC_PATH}/bzImage-to-vmlinux.sh" "${ORI_ZIMAGE_FILE}" "${TMP_PATH}/vmlinux" >"${LOG_FILE}" 2>&1 || dieLog
   # Patch boot params and ramdisk check
